chore: trivial changes

+++ tests/test-editor.php
@@ -27,22 +27,32 @@ class EditorTest extends WP_UnitTestCase {
 	private static $editor;
 
 	/**
+	 * @var bool $is_ci
+	 */
+	private static $is_ci;
+
+	/**
 	 * @SuppressWarnings(StaticAccess)
 	 */
 	public static function setUpBeforeClass() {
 		static::$app    = WP_Framework::get_instance( CHANGE_BLOCK_KEYWORDS );
 		static::$editor = Editor::get_instance( static::$app );
+		static::$is_ci  = ! empty( getenv( 'CI' ) );
 		static::reset();
 	}
 
 	public static function tearDownAfterClass() {
 		static::reset();
-		static::$app->file->delete( static::$app->define->plugin_assets_dir . DS . 'js' . DS . 'index.min.js' );
+		if ( static::$is_ci ) {
+			static::$app->file->delete( static::$app->define->plugin_assets_dir . DS . 'js' . DS . 'index.min.js' );
+		}
 	}
 
 	private static function reset() {
 		wp_dequeue_script( 'change-block-keywords' );
-		static::$app->file->put_contents( static::$app->define->plugin_assets_dir . DS . 'js' . DS . 'index.min.js', '' );
+		if ( static::$is_ci ) {
+			static::$app->file->put_contents( static::$app->define->plugin_assets_dir . DS . 'js' . DS . 'index.min.js', '' );
+		}
 	}
 
 	public function test_enqueue_block_editor_assets() {
